Updated Safecode to build with LLVM revision 166711
Numerous changes to adapt to the addition of the AddressSpace parameter to

    DataLayout::getIntPtrType(LLVMContext &C, unsigned AddressSpace).

In most cases, I was able to find a type to pass to the new

    DataLayout::tetIntPtrType(Type*)

method, thus allowing me to sidestep the issue of determining the
appropriate address space.  However, I couldn't think of a way of
obtaining a type from a Module, and thus in these cases, I passed
in the default address space (0).

Lit tests all pass, but I must admit I'm a bit nervous about this one.
I will attempt further tests on Monday.

I used the default address space in:

        StringTransform::runOnModule(Module & M)
        RegisterCustomizedAllocation::runOnModule(Module & M)
        AllocatorInfoPass::runOnModule (Module & M)