Enable gdb.mi/mi-reverse.exp after tweaking it to account for Clang's weird constant emission.
This test exercises reverse debugging, given:

do_nothing();
callme(1);

after "reverse-finish" from within the callme call, the debugger is at the
second line but under GCC's debug info it's not at the start of that line, so
"reverse-next" takes the debugger to the start of that line. Under Clang's
debug info, since the emission of the constant '1' is not associated with the
second line, the "start of line 2" and "just before the call to callme" are the
same place so "reverse-next" steps to line 1.

By using a function call to retrieve the '1' value, this collapses both the
Clang and GCC case together by ensuring "start of line 2" and "just before the
call to callme" are not the same in either case.

Especially when you look at the assembly for this case it seems rather silly
that the constants aren't associated with the line of the function call - here
Clang hasn't hoisted the constant emission far away, in all cases (even though
'2' is used multiple times) the constant emission appears immediately prior to
the use.

+16 passes